feat: Update DTO classes to use readonly properties and enhance session management with token hashing

+++ src/App/DTO/TransferDto.php
@@ -7,7 +7,7 @@ readonly class TransferDto
     public function __construct(
         public string $fromAccountId,
         public string $toAccountId,
-        public float $amount,
+        public int $amount,
         public string $categoryId,
         public ?string $description = null,
     ) {
